RavenDB-24187 : address PR comments
