+++ src/Raven.Server/Documents/Sharding/Handlers/ShardedGenAiHandler.cs
@@ -1,5 +1,4 @@
 ﻿using System.Threading.Tasks;
-using Raven.Client.Exceptions.Sharding;
 using Raven.Server.Documents.Sharding.Handlers.Processors.ETL;
 using Raven.Server.Routing;
 
@@ -12,11 +11,5 @@ public sealed class ShardedGenAiHandler : ShardedDatabaseRequestHandler
     {
         using (var processor = new ShardedGenAiHandlerProcessorForPostScriptTest(this))
             await processor.ExecuteAsync();
-    }
-
-    [RavenShardedAction("/databases/*/admin/ai/gen-ai/to-json-schema", "POST")]
-    public Task GetJsonSchemaFromSampleObject()
-    {
-        throw new NotSupportedInShardingException("GenAI is currently not supported in sharding");
     }
 }
