+++ src/Raven.Server/Web/Studio/StudioTasksHandler.cs
@@ -7,6 +7,7 @@ using Raven.Client.Util;
 using Raven.Server.Config;
 using Raven.Server.Config.Categories;
 using Raven.Server.Config.Settings;
+using Raven.Server.Documents.AI;
 using Raven.Server.Documents.ETL.Providers.ElasticSearch;
 using Raven.Server.Documents.Indexes;
 using Raven.Server.Documents.Indexes.IndexMerging;
@@ -280,6 +281,25 @@ namespace Raven.Server.Web.Studio
                 writer.WritePropertyName(nameof(NextCronExpressionOccurrence.ServerTime));
                 writer.WriteDateTime(nextOccurrence, false);
                 writer.WriteEndObject();
+            }
+        }
+
+        [RavenAction("/studio-tasks/convert-to-json-schema", "POST", AuthorizationStatus.ValidUser, EndpointType.Read)]
+        public async Task GetJsonSchemaFromSampleObject()
+        {
+            using (ServerStore.ContextPool.AllocateOperationContext(out JsonOperationContext context))
+            {
+                var sampleObj = await context.ReadForMemoryAsync(RequestBodyStream(), "convert-to-json-schema");
+
+                await using (var writer = new AsyncBlittableJsonTextWriter(context, ResponseBodyStream()))
+                {
+                    var schema = AbstractChatCompletionClient<JsonOperationContext>.GetSchemaFor(sampleObj.ToString());
+
+                    writer.WriteStartObject();
+                    writer.WritePropertyName("Result");
+                    writer.WriteString(schema);
+                    writer.WriteEndObject();
+                }
             }
         }
